Issue-142 - Add failing router TLS test
