+++ extension/src/router_channel.rs
@@ -157,6 +157,7 @@ impl RouterChannel {
     let (mut router_tx, router_recv) = mpsc::channel::<Result<Frame<Bytes>>>(32 * 1024);
     let boxed_body = BodyExt::boxed(StreamBody::new(router_recv));
     let router_request = Request::builder()
+      .version(hyper::Version::HTTP_2)
       .uri(&self.channel_url)
       .method("POST")
       .header(hyper::header::DATE, fmt_http_date(SystemTime::now()))
@@ -486,40 +487,89 @@ mod tests {
   use tokio_test::assert_ok;
 
   #[tokio::test]
-  async fn test_channel_status_305() {
-    test_channel_status_code(
+  async fn test_channel_status_305_http() {
+    fixture_channel_status_code(
       StatusCode::USE_PROXY,
       Some(ChannelResult::RouterStatusOther),
+      test_mock_router::test_mock_router::ListenerType::Http,
     )
     .await;
   }
 
   #[tokio::test]
-  async fn test_channel_status_400() {
-    test_channel_status_code(
+  async fn test_channel_status_305_https() {
+    fixture_channel_status_code(
+      StatusCode::USE_PROXY,
+      Some(ChannelResult::RouterStatusOther),
+      test_mock_router::test_mock_router::ListenerType::Https,
+    )
+    .await;
+  }
+
+  #[tokio::test]
+  async fn test_channel_status_400_http() {
+    fixture_channel_status_code(
       StatusCode::BAD_REQUEST,
       Some(ChannelResult::RouterStatus4xx),
+      test_mock_router::test_mock_router::ListenerType::Http,
     )
     .await;
   }
 
   #[tokio::test]
-  async fn test_channel_status_409() {
-    test_channel_status_code(StatusCode::CONFLICT, Some(ChannelResult::GoAwayReceived)).await;
+  async fn test_channel_status_400_https() {
+    fixture_channel_status_code(
+      StatusCode::BAD_REQUEST,
+      Some(ChannelResult::RouterStatus4xx),
+      test_mock_router::test_mock_router::ListenerType::Https,
+    )
+    .await;
   }
 
   #[tokio::test]
-  async fn test_channel_status_500() {
-    test_channel_status_code(
+  async fn test_channel_status_409_http() {
+    fixture_channel_status_code(
+      StatusCode::CONFLICT,
+      Some(ChannelResult::GoAwayReceived),
+      test_mock_router::test_mock_router::ListenerType::Http,
+    )
+    .await;
+  }
+
+  #[tokio::test]
+  async fn test_channel_status_409_https() {
+    fixture_channel_status_code(
+      StatusCode::CONFLICT,
+      Some(ChannelResult::GoAwayReceived),
+      test_mock_router::test_mock_router::ListenerType::Https,
+    )
+    .await;
+  }
+
+  #[tokio::test]
+  async fn test_channel_status_500_http() {
+    fixture_channel_status_code(
       StatusCode::INTERNAL_SERVER_ERROR,
       Some(ChannelResult::RouterStatus5xx),
+      test_mock_router::test_mock_router::ListenerType::Http,
     )
     .await;
   }
 
-  async fn test_channel_status_code(
+  #[tokio::test]
+  async fn test_channel_status_500_https() {
+    fixture_channel_status_code(
+      StatusCode::INTERNAL_SERVER_ERROR,
+      Some(ChannelResult::RouterStatus5xx),
+      test_mock_router::test_mock_router::ListenerType::Https,
+    )
+    .await;
+  }
+
+  async fn fixture_channel_status_code(
     status_code: StatusCode,
     expected_result: Option<ChannelResult>,
+    listener_type: test_mock_router::test_mock_router::ListenerType,
   ) {
     let lambda_id = "lambda_id".to_string();
     let pool_id = "pool_id".to_string();
@@ -535,6 +585,7 @@ mod tests {
         channel_panic_request_to_extension_after_start: false,
         channel_panic_request_to_extension_before_close: false,
         ping_panic_after_count: -1,
+        listener_type,
       },
     );
 
@@ -606,7 +657,24 @@ mod tests {
   }
 
   #[tokio::test]
-  async fn test_channel_read_request_send_to_app() {
+  async fn test_channel_read_request_send_to_app_http() {
+    fixture_channel_read_request_send_to_app(
+      test_mock_router::test_mock_router::ListenerType::Http,
+    )
+    .await;
+  }
+
+  #[tokio::test]
+  async fn test_channel_read_request_send_to_app_https() {
+    fixture_channel_read_request_send_to_app(
+      test_mock_router::test_mock_router::ListenerType::Https,
+    )
+    .await;
+  }
+
+  async fn fixture_channel_read_request_send_to_app(
+    listener_type: test_mock_router::test_mock_router::ListenerType,
+  ) {
     let lambda_id = "lambda_id".to_string();
     let pool_id = "pool_id".to_string();
     let channel_id = "channel_id".to_string();
@@ -622,6 +690,7 @@ mod tests {
         channel_panic_request_to_extension_after_start: false,
         channel_panic_request_to_extension_before_close: false,
         ping_panic_after_count: 0,
+        listener_type,
       },
     );
 
@@ -729,7 +798,18 @@ mod tests {
   }
 
   #[tokio::test]
-  async fn test_channel_goaway_on_body() {
+  async fn test_channel_goaway_on_body_http() {
+    fixture_channel_goaway_on_body(test_mock_router::test_mock_router::ListenerType::Http).await;
+  }
+
+  #[tokio::test]
+  async fn test_channel_goaway_on_body_https() {
+    fixture_channel_goaway_on_body(test_mock_router::test_mock_router::ListenerType::Https).await;
+  }
+
+  async fn fixture_channel_goaway_on_body(
+    listener_type: test_mock_router::test_mock_router::ListenerType,
+  ) {
     let lambda_id = "lambda_id".to_string();
     let pool_id = "pool_id".to_string();
     let channel_id = "channel_id".to_string();
@@ -745,6 +825,7 @@ mod tests {
         channel_panic_request_to_extension_after_start: false,
         channel_panic_request_to_extension_before_close: false,
         ping_panic_after_count: 0,
+        listener_type,
       },
     );
     let router_endpoint: Endpoint =
@@ -855,6 +936,7 @@ mod tests {
         channel_panic_request_to_extension_after_start: false,
         channel_panic_request_to_extension_before_close: false,
         ping_panic_after_count: 0,
+        listener_type: test_mock_router::test_mock_router::ListenerType::Http,
       },
     );
     let router_endpoint: Endpoint =
